Refactor BigDeep policy head and action indexing

+++ learning/encoding/game_encoder.pyx
@@ -68,17 +68,17 @@ cpdef tuple get_terminal_value(Armada game):
     game_length[game.round - 1] = 1.0
 
     raw_point = np.array([game.get_point(1), game.get_point(-1)], dtype=np.float32) / 200.0
-    
+
 
     return game.winner, {'game_length': game_length, 'ship_hulls': ship_hulls, 'raw_point': raw_point}
 
 
-cpdef tuple encode_game_state(Armada game, 
-                              float[:] scalar_buffer, 
-                              float[:, :] ship_entity_buffer, 
-                              float[:, :] ship_coords_buffer, 
-                              float[:, :, :] ship_def_token_buffer, 
-                              cnp.uint8_t[:, :, :, :] spatial_buffer, 
+cpdef tuple encode_game_state(Armada game,
+                              float[:] scalar_buffer,
+                              float[:, :] ship_entity_buffer,
+                              float[:, :] ship_coords_buffer,
+                              float[:, :, :] ship_def_token_buffer,
+                              cnp.uint8_t[:, :, :, :] spatial_buffer,
                               float[:, :, :] relation_buffer):
     """
     Writes encoding directly into the provided memory views.
@@ -105,43 +105,43 @@ cdef void encode_scalar_features(Armada game, float[:] scalar_view):
     cdef int offset = 0
     cdef AttackInfo attack_info
 
-    cdef tuple pool_result 
+    cdef tuple pool_result
 
     scalar_view[:] = 0.0
 
     # --- Base Game State Features ---
-    
+
     # Simple scalar values (3 features)
     scalar_view[0] = game.round / 6.0
     scalar_view[1] = game.get_point(1) / 200.0
     scalar_view[2] = game.get_point(-1) / 200.0
     offset = 3
-    
+
     # One-hot encoded initiative (2 features) (This information is NOT necessary but provides faction context)
     scalar_view[offset + (game.first_faction != 1)] = 1.0; offset += 2
     scalar_view[offset + (game.second_faction != 1)] = 1.0; offset += 2
-    
+
     # One-hot encoded phase (phase_type features)
     scalar_view[offset + <int>game.phase] = 1.0
     offset += phase_type # phase_type = 10
-    
+
     # One-hot encoded current player (2 features)
     scalar_view[offset + (game.current_player != 1)] = 1.0
     offset += 2
-    
+
     # --- Attack Context Features (17 features) ---
-    
+
     if game.attack_info is None:
         return # No attack in progress, return early
-        
+
     attack_info = game.attack_info
-    
+
     # Attack availability flags (3 features)
     scalar_view[offset] = float(attack_info.con_fire_dial)
     scalar_view[offset + 1] = float(attack_info.con_fire_token)
     scalar_view[offset + 2] = float(attack_info.swarm)
     offset += 3
-    
+
     # Dice Pool Result (11 features)
     # Black (3)
     pool_result = attack_info.attack_pool_result[Dice.BLACK]
@@ -149,14 +149,14 @@ cdef void encode_scalar_features(Armada game, float[:] scalar_view):
     scalar_view[offset + 1] = <float>pool_result[1]
     scalar_view[offset + 2] = <float>pool_result[2]
     offset += 3
-    
+
     # Blue (3)
     pool_result = attack_info.attack_pool_result[Dice.BLUE]
     scalar_view[offset] = <float>pool_result[0]
     scalar_view[offset + 1] = <float>pool_result[1]
     scalar_view[offset + 2] = <float>pool_result[2]
     offset += 3
-    
+
     # Red (5)
     pool_result = attack_info.attack_pool_result[Dice.RED]
     scalar_view[offset] = <float>pool_result[0]
@@ -165,7 +165,7 @@ cdef void encode_scalar_features(Armada game, float[:] scalar_view):
     scalar_view[offset + 3] = <float>pool_result[3]
     scalar_view[offset + 4] = <float>pool_result[4]
     offset += 5
-    
+
     # Critical Effect (one-hot, 1 feature)
     if attack_info.critical is not None:
         scalar_view[offset + <int>attack_info.critical] = 1.0
@@ -175,27 +175,27 @@ cdef void encode_scalar_features(Armada game, float[:] scalar_view):
     scalar_view[offset] = (<int>attack_info.attack_range) / 4.0
     scalar_view[offset + 1] = 1.0 if attack_info.obstructed else 0.0
     offset += 2 # End of array
-    
+
 
 
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-cdef void encode_ship_entity_features(Armada game, 
-                                      float[:, :] ship_view_buffer, 
-                                      float[:, :] coords_view_buffer, 
+cdef void encode_ship_entity_features(Armada game,
+                                      float[:, :] ship_view_buffer,
+                                      float[:, :] coords_view_buffer,
                                       float[:, :, :] def_token_view_buffer):
     """
     Encodes a detailed vector for each ship, now including its role in an active attack.
-    
-    This optimized version writes directly into the pre-allocated 
+
+    This optimized version writes directly into the pre-allocated
     game.ship_encode_array to avoid memory allocations within the loop.
     """
     # --- C-level variable definitions ---
     cdef Ship ship
     cdef DefenseToken token
     cdef AttackInfo attack_info
-    
+
     # Memory view for the target row
     cdef cnp.float32_t[:] single_ship_view
     cdef cnp.float32_t[:] single_coords_view
@@ -203,7 +203,7 @@ cdef void encode_ship_entity_features(Armada game,
     cdef float[:, :] static_stats_view = game.ship_static_encode_array
     # Indices and iterators
     cdef int offset, stack_idx, defense_start_idx, hull, command
-    
+
     # Booleans
     cdef bint is_attack = False
 
@@ -219,7 +219,7 @@ cdef void encode_ship_entity_features(Armada game,
     if game.attack_info is not None:
         attack_info = game.attack_info
         is_attack = True
-    
+
     for ship in game.ships:
         if ship.id >= max_ships or ship.destroyed: continue
 
@@ -243,7 +243,7 @@ cdef void encode_ship_entity_features(Armada game,
 
             if not(is_attack and (token.id in attack_info.spent_token_indices or token.type in attack_info.spent_token_types)):
                 single_def_token_view[3] = 1.0
-            
+
             single_def_token_view[<int>token.type + 4] = 1.0  # One-hot encoding of token type
 
 
@@ -266,13 +266,13 @@ cdef void encode_ship_entity_features(Armada game,
         for hull in range(c_hull_type):
             single_ship_view[offset + hull] = ship.shield[hull] / global_max_shields
         offset += c_hull_type
-        
+
         # --- Command Stack (12 features) ---
         if ship.team == game.simulation_player:
             for stack_idx, command in enumerate(ship.command_stack):
                 single_ship_view[offset + stack_idx * c_command_type + command] = 1.0
         offset += max_command_stack * c_command_type # Advance offset by the block size
-        
+
         # --- Command Dials (4 features) ---
         for command in ship.command_dial:
             single_ship_view[offset + command] = 1.0
@@ -302,7 +302,7 @@ cdef void encode_ship_entity_features(Armada game,
 #     Encodes a detailed vector for each squad in-place in game.squad_encode_array
 #     to avoid all memory allocations, then returns a reference to the array.
 #     """
-    
+
 #     cdef Squad squad
 #     cdef cnp.float32_t[:] squad_view
 #     cdef int defense_idx, offset, overlap_start_idx, defense_start_idx
@@ -320,7 +320,7 @@ cdef void encode_ship_entity_features(Armada game,
 #     for squad in game.squads:
 #         if squad.id >= max_squads or squad.destroyed:
 #             continue
-            
+
 #         # Get a view to the specific row we will write to
 #         squad_view = game.squad_encode_array[squad.id]
 #         offset = SQUAD_STATS_FEATURES
@@ -332,7 +332,7 @@ cdef void encode_ship_entity_features(Armada game,
 #         squad_view[offset] = squad.can_move; offset += 1
 #         squad_view[offset] = squad.coords[0] / game.player_edge; offset += 1
 #         squad_view[offset] = squad.coords[1] / game.short_edge; offset += 1
-        
+
 #         # --- Overlap (MAX_SHIPS=6 features) ---
 #         overlap_start_idx = offset
 #         if squad.overlap_ship_id is not None:
@@ -346,7 +346,7 @@ cdef void encode_ship_entity_features(Armada game,
 #             if not attack_info.is_defender_ship and squad.id == attack_info.defend_squad_id:
 #                 squad_view[offset + 1] = 1.0
 #         offset += 2
-        
+
 #         # # --- Defense Token (2*4 = 8 features) ---
 #         # defense_start_idx = offset
 #         # for defense_idx, token in squad.defense_tokens.items():
@@ -370,31 +370,31 @@ cdef void encode_spatial_mask(Armada game, cnp.uint8_t[:, :, :, :] planes_view):
     Fills the game.spatial_planes 2D grid representations of the game board
     in-place and returns a reference to it.
     """
-    
+
     # Standard cleanup
     planes_view[:] = 0
 
     cdef int i, r, c
-    cdef long[:] rr_view, cc_view 
+    cdef long long[:] rr_view, cc_view
     cdef int hull, attack_range, channel_idx
     cdef dict threat_plane_dict, ranges_dict
     cdef Ship ship
     cdef tuple ship_hash
-    
+
     for ship in game.ships:
         if ship.id >= Config.MAX_SHIPS or ship.destroyed:
             continue
-            
+
         ship_hash = ship.get_ship_hash_state()
 
         # --- 1. Ship Presence ---
         # Get arrays as usual
         rr, cc = cache._ship_presence_indices(ship_hash)
-        
+
         # Cast to memoryviews for C-level access
         rr_view = rr
         cc_view = cc
-        
+
         # C-Level Loop (No GIL, No Python Overhead)
         for i in range(rr_view.shape[0]):
             r = rr_view[i]
@@ -405,17 +405,17 @@ cdef void encode_spatial_mask(Armada game, cnp.uint8_t[:, :, :, :] planes_view):
 
         # --- 2. Ship Threat ---
         threat_plane_dict = cache._ship_threat_indices(ship_hash)
-        
+
         # Iterate Dictionary Items to avoid KeyError risks
         for hull, ranges_dict in threat_plane_dict.items():
             for attack_range, coords in ranges_dict.items():
-                
+
                 # Calculate channel carefully
                 channel_idx = 1 + (hull * 3) + attack_range
-                
+
                 rr_view = coords[0]
                 cc_view = coords[1]
-                
+
                 for i in range(rr_view.shape[0]):
                     r = rr_view[i]
                     c = cc_view[i]
@@ -429,15 +429,15 @@ cdef void encode_relation_matrix(Armada game, float[:, :, :] rel_matrix):
     """
     Optimized pairwise range encoding.
     """
-    
+
     cdef int i, j, attacker_id, defender_id
     cdef int from_hull, to_hull
     cdef int flat_idx
     cdef float attack_range, dx, dy, dist, relative_bearing
-    cdef list ships = game.ships 
+    cdef list ships = game.ships
     cdef int n_ships = len(ships)
-    cdef Ship attacker, defender 
-    
+    cdef Ship attacker, defender
+
     rel_matrix[:] = 0.0
 
     cdef list range_list, attack_range_list
@@ -445,50 +445,50 @@ cdef void encode_relation_matrix(Armada game, float[:, :, :] rel_matrix):
     # --- MAIN LOOP ---
     for i in range(n_ships):
         attacker = ships[i]
-        
-        # Check destroyed status 
+
+        # Check destroyed status
         if attacker.destroyed:
             continue
-            
+
         attacker_id = attacker.id
-        
+
         for j in range(n_ships):
             # Self-check optimization
-            if i == j: 
+            if i == j:
                 continue
 
             defender = ships[j]
-            
+
             if defender.destroyed:
                 continue
-                
+
             defender_id = defender.id
 
             # Attack Range Ralation
             _, range_list = cache.attack_range_s2s(
-                attacker.get_ship_hash_state(), 
+                attacker.get_ship_hash_state(),
                 defender.get_ship_hash_state()
             )
 
             for from_hull in range(c_hull_type):
-                attack_range_list = range_list[from_hull] 
-                
+                attack_range_list = range_list[from_hull]
+
                 for to_hull in range(c_hull_type):
                     attack_range = <float>attack_range_list[to_hull]
-                    
+
                     flat_idx = from_hull * c_hull_type + to_hull
-                    
+
                     rel_matrix[attacker_id, defender_id, flat_idx] = attack_range
 
             # Standard Geometric Relation
             dx = (defender.x - attacker.x) / game.player_edge
             dy = (defender.y - attacker.y) / game.short_edge
             dist = sqrt(dx*dx + dy*dy)
-            
+
             # Angle relative to attacker's front (Bearing)
             angle_to_target = atan2(dy, dx)
             relative_bearing = angle_to_target - attacker.orientation
-            
+
             # --- Store them at the end of the vector ---
             # Indices 0-15 are Hull Ranges
             # Indices 16-19 are Relative Physics
